cmd/tailscale/cli,clientupdate: extract new clientupdate package
Extract the self-update logic from cmd/tailscale/cli into a standalone
package that could be used from tailscaled later.

Updates #6995
